Move other stuff in React category panel and scene file
